model/requirements.txt
onnxruntime>=1.15.0
numpy>=1.21.0
Pillow>=9.0.0
pyyaml>=6.0.0
